+++ src/components/ui/Button/DeleteButton.tsx
@@ -1,10 +1,7 @@
 import type { VFC } from "react";
-import { useRecoilState } from "recoil";
 
 import { TrashIcon } from "~/components/ui/Assets/HeroIcon";
-import type { TodoListState } from "~/globalStates/atoms/todoListState";
-import { todoListState } from "~/globalStates/atoms/todoListState";
-import { useDeleteTodoMutation } from "$/gql";
+import { useDeleteTodo } from "~/hooks";
 
 type DeleteButtonProps = {
   className?: string;
@@ -12,26 +9,7 @@ type DeleteButtonProps = {
 };
 
 export const DeleteButton: VFC<DeleteButtonProps> = ({ className, todoId }) => {
-  const [deleteTodo] = useDeleteTodoMutation();
-  const [todoList, setTodoList] = useRecoilState<TodoListState>(todoListState);
-
-  const handleDeleteTodo = async () => {
-    try {
-      const { data } = await deleteTodo({
-        variables: {
-          input: {
-            todoId: todoId,
-          },
-        },
-      });
-      if (data && data.deleteTodo) {
-        const newTodoList = todoList.filter((t) => t && t.id !== todoId);
-        setTodoList(newTodoList);
-      }
-    } catch (e) {
-      console.error(`${e}: Todoの削除に失敗しました`);
-    }
-  };
+  const handleDeleteTodo = useDeleteTodo(todoId);
 
   return (
     <button className={className} onClick={handleDeleteTodo}>
