+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useAuth } from "./useAuth";
+export { useDeleteTodo } from "./useDeleteTodo";
